+++ CHANGELOG.rst
@@ -1,5 +1,6 @@
 Development
 ===========================
+* Asynchronous Redis locking primitive for distributing computing
 * Added the :ref:`Twitter Streaming <tutorials-tweets>` tutorial
 * Added Javascript directory in examples and a gruntfile for compiling and linting scripts
 * Better handling of Ctrl-C in the test application